Move CondLink to core
