+++ opencog/atoms/core/CondLink.h
